+++ src/main/java/com/jrzln/mibalanceapi/wallets/domain/model/commands/AddBalanceToWalletCommand.java
@@ -1,6 +1,8 @@
 package com.jrzln.mibalanceapi.wallets.domain.model.commands;
 
+import java.math.BigDecimal;
+
 /**
  * Command to add balance to a wallet.
  */
-public record AddBalanceToWalletCommand(String walletId, double amountToAdd) {}
+public record AddBalanceToWalletCommand(String walletId, BigDecimal amountToAdd) {}
